[IFC][Integration] Move box geometry update to a dedicate class https://bugs.webkit.org/show_bug.cgi?id=265387
Reviewed by Antti Koivisto.

This is in preparation for supporting intrinsic width computation on inline level boxes.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/layout/integration/LayoutIntegrationBoxTree.h:
* Source/WebCore/layout/integration/inline/LayoutIntegrationBoxGeometryUpdater.cpp: Added.
(WebCore::LayoutIntegration::BoxGeometryUpdater::BoxGeometryUpdater):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateReplacedDimensions):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateInlineBlockDimensions):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateInlineTableDimensions):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateListItemDimensions):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateListMarkerDimensions):
(WebCore::LayoutIntegration::contentLogicalWidthForRenderer):
(WebCore::LayoutIntegration::contentLogicalHeightForRenderer):
(WebCore::LayoutIntegration::horizontalLogicalMargin):
(WebCore::LayoutIntegration::verticalLogicalMargin):
(WebCore::LayoutIntegration::logicalBorder):
(WebCore::LayoutIntegration::logicalPadding):
(WebCore::LayoutIntegration::scrollbarLogicalSize):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateLayoutBoxDimensions):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateLineBreakBoxDimensions):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateInlineBoxDimensions):
(WebCore::LayoutIntegration::BoxGeometryUpdater::setGeometriesForLayout):
(WebCore::LayoutIntegration::BoxGeometryUpdater::updateInlineContentConstraints):
(WebCore::LayoutIntegration::BoxGeometryUpdater::rootLayoutBox const):
(WebCore::LayoutIntegration::BoxGeometryUpdater::rootLayoutBox):
* Source/WebCore/layout/integration/inline/LayoutIntegrationBoxGeometryUpdater.h: Added.
(WebCore::LayoutIntegration::BoxGeometryUpdater::takeNestedListMarkerOffsets):
(WebCore::LayoutIntegration::BoxGeometryUpdater::boxTree):
(WebCore::LayoutIntegration::BoxGeometryUpdater::boxTree const):
(WebCore::LayoutIntegration::BoxGeometryUpdater::layoutState):
(WebCore::LayoutIntegration::BoxGeometryUpdater::layoutState const):
* Source/WebCore/layout/integration/inline/LayoutIntegrationLineLayout.cpp:
(WebCore::LayoutIntegration::LineLayout::LineLayout):
(WebCore::LayoutIntegration::LineLayout::updateInlineContentDimensions):
(WebCore::LayoutIntegration::LineLayout::layout):
(WebCore::LayoutIntegration::LineLayout::updateInlineContentConstraints):
(WebCore::LayoutIntegration::LineLayout::updateReplacedDimensions): Deleted.
(WebCore::LayoutIntegration::LineLayout::updateInlineBlockDimensions): Deleted.
(WebCore::LayoutIntegration::LineLayout::updateInlineTableDimensions): Deleted.
(WebCore::LayoutIntegration::LineLayout::updateListItemDimensions): Deleted.
(WebCore::LayoutIntegration::LineLayout::updateListMarkerDimensions): Deleted.
(WebCore::LayoutIntegration::contentLogicalWidthForRenderer): Deleted.
(WebCore::LayoutIntegration::contentLogicalHeightForRenderer): Deleted.
(WebCore::LayoutIntegration::horizontalLogicalMargin): Deleted.
(WebCore::LayoutIntegration::verticalLogicalMargin): Deleted.
(WebCore::LayoutIntegration::logicalBorder): Deleted.
(WebCore::LayoutIntegration::logicalPadding): Deleted.
(WebCore::LayoutIntegration::scrollbarLogicalSize): Deleted.
(WebCore::LayoutIntegration::LineLayout::updateLayoutBoxDimensions): Deleted.
(WebCore::LayoutIntegration::LineLayout::updateLineBreakBoxDimensions): Deleted.
(WebCore::LayoutIntegration::LineLayout::updateInlineBoxDimensions): Deleted.
(WebCore::LayoutIntegration::LineLayout::prepareLayoutState): Deleted.
* Source/WebCore/layout/integration/inline/LayoutIntegrationLineLayout.h:

Canonical link: https://commits.webkit.org/271156@main
